backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-multipart==0.0.6
Pillow==10.1.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
boto3==1.34.10
cryptography==41.0.7
requests==2.31.0
pytz==2023.3
timezonefinder==6.2.0

